requirements.txt
discord
python-dotenv
